Python requirements.txt
flask==2.3.3
numpy==1.25.2
faiss-cpu==1.7.4
requests==2.31.0
werkzeug==2.3.7